requirements.txt
playwright==1.56.0
pytest==8.4.2
pytest-playwright==0.7.1
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-xdist==3.8.0
pytest-bdd==8.1.0
pytest-base-url==2.1.0
requests==2.32.5
allure-pytest==2.13.5
faker==40.15.0